refactor: move hook scripts into hooks/guards/ and hooks/formatters/
Move 18 guard/gate/advisor scripts to guards/ and 5 formatter modules
to formatters/. Fix sys.path bootstraps for cross-directory bare imports.
Update hooks/__init__.py package-style imports.
