+++ src/autoskillit/hooks/__init__.py
@@ -5,12 +5,12 @@ from autoskillit.hook_registry import (
     HookDef,
     generate_hooks_json,
 )
-from autoskillit.hooks._fmt_primitives import _HOOK_CONFIG_PATH_COMPONENTS
-from autoskillit.hooks.branch_protection_guard import BRANCH_PROTECTION_DENY_TRIGGER
-from autoskillit.hooks.leaf_orchestration_guard import LEAF_ORCHESTRATION_DENY_TRIGGER
-from autoskillit.hooks.quota_guard import QUOTA_GUARD_DENY_TRIGGER
+from autoskillit.hooks.formatters._fmt_primitives import _HOOK_CONFIG_PATH_COMPONENTS
+from autoskillit.hooks.guards.branch_protection_guard import BRANCH_PROTECTION_DENY_TRIGGER
+from autoskillit.hooks.guards.leaf_orchestration_guard import LEAF_ORCHESTRATION_DENY_TRIGGER
+from autoskillit.hooks.guards.quota_guard import QUOTA_GUARD_DENY_TRIGGER
 from autoskillit.hooks.quota_post_hook import QUOTA_POST_WARNING_TRIGGER
-from autoskillit.hooks.review_loop_gate import REVIEW_LOOP_DENY_TRIGGER
+from autoskillit.hooks.guards.review_loop_gate import REVIEW_LOOP_DENY_TRIGGER
 
 __all__ = [
     "HOOK_REGISTRY",
