+++ src/autoskillit/hooks/formatters/__init__.py
@@ -1,0 +1,1 @@
+"""PostToolUse formatter scripts — MCP JSON to Markdown-KV reformatter."""
