+++ src/autoskillit/hooks/formatters/_fmt_execution.py
