+++ src/autoskillit/hooks/formatters/_fmt_status.py
