+++ src/autoskillit/hooks/guards/__init__.py
@@ -1,0 +1,1 @@
+"""PreToolUse guard scripts — invoked as standalone python3 commands."""
