+++ src/autoskillit/hooks/guards/generated_file_write_guard.py
