+++ src/autoskillit/hooks/guards/grep_pattern_lint_guard.py
