+++ src/autoskillit/hooks/guards/mcp_health_guard.py
