+++ src/autoskillit/hooks/guards/open_kitchen_guard.py
