+++ src/autoskillit/hooks/guards/skill_cmd_guard.py
