+++ src/autoskillit/hooks/guards/skill_command_guard.py
