+++ src/autoskillit/hooks/guards/unsafe_install_guard.py
