+++ src/autoskillit/hooks/guards/write_guard.py
